- Main Update: When a GetOneById method in either WatcherGenresManager or WatcherTitlesManager would return null, it calls the Create method to insert a new entry in the database with default values and the corresponding ids, then it calls the GenericRepository's GetOneById method again and returns the result (which should be the new entry) (see #46, #45). - Refactoring: Introduced try catch in GenericRepository's InsertAsync method. - Refactoring: Update and delete methods in GenericRepository are now Task<bool> and their counterparts in managers now handle the null pointer returned values (see #26). - Refactoring: Added several Console.Writelines to warn about retrieving null pointers in FillModelsLists methods in WatcherGenresManager and WatcherTitlesManager. - Refactoring: Changed inconsistent explicit types to implicit in managers (where I considered it is not useful having the explicit type displayed).

+++ Movie4U/Repositories/GenericRepository.cs
@@ -189,9 +189,17 @@ namespace Movie4U.Repositories
 
         public virtual async Task<TEntity> InsertAsync(TEntity entity)
         {
-            await entities.AddAsync(entity);
-            await db.SaveChangesAsync();
-            return entity;
+            try
+            {
+                await entities.AddAsync(entity);
+                await db.SaveChangesAsync();
+                return entity;
+            }
+            catch(Exception e)
+            {
+                Console.WriteLine(e.ToString());
+                return null;
+            }
         }
 
         public async Task<TEntity[]> InsertMultipleAsync(TEntity[] entities)
@@ -208,26 +216,35 @@ namespace Movie4U.Repositories
             return entities;
         }
 
-        public virtual async Task UpdateAsync(TEntity entityToUpdate)
+        public virtual async Task<bool> UpdateAsync(TEntity entityToUpdate)
         {
             entities.Attach(entityToUpdate);
             db.Entry(entityToUpdate).State = EntityState.Modified;
             await db.SaveChangesAsync();
+
+            return true;
         }
 
-        public virtual async Task DeleteAsync(object id)
+        public virtual async Task<bool> DeleteAsync(object id)
         {
             var entityToDelete = await entities.FindAsync(id);
-            await DeleteAsync(entityToDelete);
+            if (entityToDelete == null)
+                return false;
+
+            var result = await DeleteAsync(entityToDelete);
             await db.SaveChangesAsync();
+
+            return result;
         }
 
-        public virtual async Task DeleteAsync(TEntity entityToDelete)
+        public virtual async Task<bool> DeleteAsync(TEntity entityToDelete)
         {
             if (db.Entry(entityToDelete).State == EntityState.Detached)
                 entities.Attach(entityToDelete);
             entities.Remove(entityToDelete);
             await db.SaveChangesAsync();
+
+            return true;
         }
